Refactor and improve blockchain data client
