+++ crates/adapters/blockchain/src/data/subscription.rs
@@ -20,7 +20,7 @@ use alloy::primitives::{Address, keccak256};
 use nautilus_model::defi::DexType;
 
 /// Manages subscriptions to DeFi protocol events (swaps, mints, burns) across different DEXs.
-/// 
+///
 /// This manager tracks which pool addresses are subscribed for each event type
 /// and maintains the event signature encodings for efficient filtering.
 #[derive(Debug)]
@@ -214,33 +214,69 @@ mod tests {
 
     #[rstest]
     fn test_new_creates_empty_manager(manager: DefiDataSubscriptionManager) {
-        assert_eq!(manager.get_subscribed_dex_contract_addresses(&DexType::UniswapV3).len(), 0);
-        assert_eq!(manager.get_subscribed_dex_event_signatures(&DexType::UniswapV3).len(), 0);
-        assert!(manager.get_dex_pool_swap_event_signature(&DexType::UniswapV3).is_none());
-        assert!(manager.get_dex_pool_mint_event_signature(&DexType::UniswapV3).is_none());
-        assert!(manager.get_dex_pool_burn_event_signature(&DexType::UniswapV3).is_none());
+        assert_eq!(
+            manager
+                .get_subscribed_dex_contract_addresses(&DexType::UniswapV3)
+                .len(),
+            0
+        );
+        assert_eq!(
+            manager
+                .get_subscribed_dex_event_signatures(&DexType::UniswapV3)
+                .len(),
+            0
+        );
+        assert!(
+            manager
+                .get_dex_pool_swap_event_signature(&DexType::UniswapV3)
+                .is_none()
+        );
+        assert!(
+            manager
+                .get_dex_pool_mint_event_signature(&DexType::UniswapV3)
+                .is_none()
+        );
+        assert!(
+            manager
+                .get_dex_pool_burn_event_signature(&DexType::UniswapV3)
+                .is_none()
+        );
     }
 
     #[rstest]
     fn test_register_dex_for_subscriptions(registered_manager: DefiDataSubscriptionManager) {
         // Should have all three event signatures
-        let signatures = registered_manager.get_subscribed_dex_event_signatures(&DexType::UniswapV3);
+        let signatures =
+            registered_manager.get_subscribed_dex_event_signatures(&DexType::UniswapV3);
         assert_eq!(signatures.len(), 3);
-        
+
         // Each signature should be properly encoded
-        assert!(registered_manager.get_dex_pool_swap_event_signature(&DexType::UniswapV3).is_some());
-        assert!(registered_manager.get_dex_pool_mint_event_signature(&DexType::UniswapV3).is_some());
-        assert!(registered_manager.get_dex_pool_burn_event_signature(&DexType::UniswapV3).is_some());
+        assert!(
+            registered_manager
+                .get_dex_pool_swap_event_signature(&DexType::UniswapV3)
+                .is_some()
+        );
+        assert!(
+            registered_manager
+                .get_dex_pool_mint_event_signature(&DexType::UniswapV3)
+                .is_some()
+        );
+        assert!(
+            registered_manager
+                .get_dex_pool_burn_event_signature(&DexType::UniswapV3)
+                .is_some()
+        );
     }
 
     #[rstest]
     fn test_subscribe_and_get_addresses(mut registered_manager: DefiDataSubscriptionManager) {
         let pool_address = address!("1234567890123456789012345678901234567890");
-        
+
         // Subscribe to swap events
         registered_manager.subscribe_swaps(DexType::UniswapV3, pool_address);
-        
-        let addresses = registered_manager.get_subscribed_dex_contract_addresses(&DexType::UniswapV3);
+
+        let addresses =
+            registered_manager.get_subscribed_dex_contract_addresses(&DexType::UniswapV3);
         assert_eq!(addresses.len(), 1);
         assert_eq!(addresses[0], pool_address);
     }
@@ -248,12 +284,12 @@ mod tests {
     #[rstest]
     fn test_subscribe_to_unregistered_dex(mut manager: DefiDataSubscriptionManager) {
         let pool_address = address!("1234567890123456789012345678901234567890");
-        
+
         // Try to subscribe without registering - should log warning but not panic
         manager.subscribe_swaps(DexType::UniswapV3, pool_address);
         manager.subscribe_mints(DexType::UniswapV3, pool_address);
         manager.subscribe_burns(DexType::UniswapV3, pool_address);
-        
+
         // Should return empty results
         let addresses = manager.get_subscribed_dex_contract_addresses(&DexType::UniswapV3);
         assert_eq!(addresses.len(), 0);
@@ -262,18 +298,28 @@ mod tests {
     #[rstest]
     fn test_unsubscribe_removes_address(mut registered_manager: DefiDataSubscriptionManager) {
         let pool_address = address!("1234567890123456789012345678901234567890");
-        
+
         // Subscribe
         registered_manager.subscribe_swaps(DexType::UniswapV3, pool_address);
-        
+
         // Verify subscription
-        assert_eq!(registered_manager.get_subscribed_dex_contract_addresses(&DexType::UniswapV3).len(), 1);
-        
+        assert_eq!(
+            registered_manager
+                .get_subscribed_dex_contract_addresses(&DexType::UniswapV3)
+                .len(),
+            1
+        );
+
         // Unsubscribe
         registered_manager.unsubscribe_swaps(DexType::UniswapV3, pool_address);
-        
+
         // Verify removal
-        assert_eq!(registered_manager.get_subscribed_dex_contract_addresses(&DexType::UniswapV3).len(), 0);
+        assert_eq!(
+            registered_manager
+                .get_subscribed_dex_contract_addresses(&DexType::UniswapV3)
+                .len(),
+            0
+        );
     }
 
     #[rstest]
@@ -281,7 +327,7 @@ mod tests {
         let swap_sig = registered_manager.get_dex_pool_swap_event_signature(&DexType::UniswapV3);
         let mint_sig = registered_manager.get_dex_pool_mint_event_signature(&DexType::UniswapV3);
         let burn_sig = registered_manager.get_dex_pool_burn_event_signature(&DexType::UniswapV3);
-        
+
         // All should be Some and start with 0x
         assert!(swap_sig.is_some() && swap_sig.unwrap().starts_with("0x"));
         assert!(mint_sig.is_some() && mint_sig.unwrap().starts_with("0x"));
@@ -291,29 +337,33 @@ mod tests {
     #[rstest]
     fn test_multiple_subscriptions_same_pool(mut registered_manager: DefiDataSubscriptionManager) {
         let pool_address = address!("1234567890123456789012345678901234567890");
-        
+
         // Subscribe same address multiple times to same event type
         registered_manager.subscribe_swaps(DexType::UniswapV3, pool_address);
         registered_manager.subscribe_swaps(DexType::UniswapV3, pool_address);
-        
+
         // Should only appear once (HashSet behavior)
-        let addresses = registered_manager.get_subscribed_dex_contract_addresses(&DexType::UniswapV3);
+        let addresses =
+            registered_manager.get_subscribed_dex_contract_addresses(&DexType::UniswapV3);
         assert_eq!(addresses.len(), 1);
     }
 
     #[rstest]
-    fn test_get_combined_addresses_from_all_events(mut registered_manager: DefiDataSubscriptionManager) {
+    fn test_get_combined_addresses_from_all_events(
+        mut registered_manager: DefiDataSubscriptionManager,
+    ) {
         let pool1 = address!("1111111111111111111111111111111111111111");
         let pool2 = address!("2222222222222222222222222222222222222222");
         let pool3 = address!("3333333333333333333333333333333333333333");
-        
+
         // Subscribe different pools to different events
         registered_manager.subscribe_swaps(DexType::UniswapV3, pool1);
         registered_manager.subscribe_mints(DexType::UniswapV3, pool2);
         registered_manager.subscribe_burns(DexType::UniswapV3, pool3);
-        
+
         // Should get all unique addresses
-        let addresses = registered_manager.get_subscribed_dex_contract_addresses(&DexType::UniswapV3);
+        let addresses =
+            registered_manager.get_subscribed_dex_contract_addresses(&DexType::UniswapV3);
         assert_eq!(addresses.len(), 3);
         assert!(addresses.contains(&pool1));
         assert!(addresses.contains(&pool2));
@@ -324,12 +374,14 @@ mod tests {
     fn test_event_signature_encoding(registered_manager: DefiDataSubscriptionManager) {
         // Known event signature and its expected keccak256 hash
         // Swap(address,address,int256,int256,uint160,uint128,int24) for UniswapV3
-        let swap_sig = registered_manager.get_dex_pool_swap_event_signature(&DexType::UniswapV3).unwrap();
-        
+        let swap_sig = registered_manager
+            .get_dex_pool_swap_event_signature(&DexType::UniswapV3)
+            .unwrap();
+
         // Should be properly formatted hex string
         assert!(swap_sig.starts_with("0x"));
         assert_eq!(swap_sig.len(), 66); // 0x + 64 hex chars (32 bytes)
-        
+
         // Verify it's valid hex
         let hex_part = &swap_sig[2..];
         assert!(hex_part.chars().all(|c| c.is_ascii_hexdigit()));
@@ -342,7 +394,7 @@ mod tests {
         let mut manager = DefiDataSubscriptionManager::new();
         let pool1 = address!("1111111111111111111111111111111111111111");
         let pool2 = address!("2222222222222222222222222222222222222222");
-        
+
         // Step 1: Register DEX
         manager.register_dex_for_subscriptions(
             dex_type,
@@ -350,27 +402,27 @@ mod tests {
             "Mint(address,uint256)",
             "Burn(address,uint256)",
         );
-        
+
         // Step 2: Subscribe to events
         manager.subscribe_swaps(dex_type, pool1);
         manager.subscribe_swaps(dex_type, pool2);
         manager.subscribe_mints(dex_type, pool1);
         manager.subscribe_burns(dex_type, pool2);
-        
+
         // Step 3: Verify subscriptions
         let addresses = manager.get_subscribed_dex_contract_addresses(&dex_type);
         assert_eq!(addresses.len(), 2);
         assert!(addresses.contains(&pool1));
         assert!(addresses.contains(&pool2));
-        
+
         // Step 4: Get event signatures
         let signatures = manager.get_subscribed_dex_event_signatures(&dex_type);
         assert_eq!(signatures.len(), 3);
-        
+
         // Step 5: Unsubscribe from some events
         manager.unsubscribe_swaps(dex_type, pool1);
         manager.unsubscribe_burns(dex_type, pool2);
-        
+
         // Step 6: Verify remaining subscriptions (only pool1 mint remains)
         let remaining = manager.get_subscribed_dex_contract_addresses(&dex_type);
         assert!(remaining.contains(&pool1)); // Still has mint subscription
